Add X-AppImage-Version to generated AppImages
This was suggested by @TheAssassin in #1027; it serves to allow
AppImages of different versions of the same software to be distinguished
by helper applications.

I also took the opportunity to move the AppImage generation into its own
separate script to hopefully make future work on our AppImages a little
easier. It is also for that reason that I put slightly more effort into
the script than what would have been needed to just make it work with
Travis.

Also as part of this change, I modified our Travis script to use an
actual version description in the generated artifact file names for
builds from the release branch. In addition to making the file names
more meaningful, this also serves the important purpose of visibly
indicating when the version number included in the AppImage is
incorrect, as it is taken from git tags and thus requires an
appropriate tag to be set on the current commit at build time. For
consistency, however, this change has been applied to builds for all
platforms.